[Fix] migration files

+++ db/migrate/20200905091106_create_likes.rb
@@ -2,7 +2,7 @@ class CreateLikes < ActiveRecord::Migration[5.2]
   def change
     create_table :likes do |t|
       t.references :user, type: :integer, foreign_key: true
-      t.references :post, type: :bigint, foreign_key: true
+      t.references :post, type: :integer, foreign_key: true
 
       t.timestamps
     end
