+++ db/migrate/20200905091200_create_post_comments.rb
@@ -2,7 +2,7 @@ class CreatePostComments < ActiveRecord::Migration[5.2]
   def change
     create_table :post_comments do |t|
       t.references :user, type: :integer, foreign_key: true
-      t.references :post, type: :bigint, foreign_key: true
+      t.references :post, type: :integer, foreign_key: true
       t.text :comment
 
       t.timestamps
